Some Rust file renames
rename lglyph to composite per C code
rename glyph to display_traits per C code dispextern.c
