+++ rust_src/crates/emacs/src/display_traits/mod.rs
@@ -1,0 +1,9 @@
+// Interface definitions for display code.
+
+use crate::bindings::glyph;
+use crate::lisp::ExternalPtr;
+pub type GlyphRef = ExternalPtr<glyph>;
+#[cfg(feature = "window-system")]
+mod glyph_string;
+#[cfg(feature = "window-system")]
+pub use glyph_string::*;
