+++ rust_src/crates/emacs/src/lib.rs
@@ -44,15 +44,14 @@ pub mod buffer;
 #[cfg(feature = "window-system")]
 pub mod color;
 #[cfg(feature = "window-system")]
+pub mod composite;
+#[cfg(feature = "window-system")]
 pub mod display_info;
+pub mod display_traits;
 pub mod eval;
 pub mod font;
 pub mod frame;
-#[cfg(feature = "window-system")]
-pub mod glyph;
 pub mod keyboard;
-#[cfg(feature = "window-system")]
-pub mod lglyph;
 pub mod list;
 pub mod multibyte;
 pub mod number;
